+++ app/components/map/navigraph/NavigraphRoute.vue
@@ -187,7 +187,9 @@ async function update() {
                 });
             }
 
-            if (dataStore.vatspy.value?.data.keyAirports.realIcao[arrival] && !hasApproach) {
+            const hideLineIfNoProcedure = getKeyedValueFromSettings('map.navigraph.routeParsing.airportOverlay.hideLineIfNoProcedure');
+
+            if (dataStore.vatspy.value?.data.keyAirports.realIcao[arrival] && !hasApproach && (!hideLineIfNoProcedure || extendedPilot)) {
                 const lastIndex = waypoints.findIndex(x => x.kind === 'missedApproach');
                 const index = lastIndex === -1 ? waypoints.length - 1 : lastIndex;
 
@@ -287,6 +289,8 @@ async function update() {
 
             const waypointForCid = dataStore.navigraphWaypoints.value[cid.toString()];
 
+            let i = 0;
+
             const onFirstWaypoint = (identifier: string, newCoordinate: Coordinate, kind: NavigraphNavDataEnrouteWaypointPartial['kind']) => {
                 if (firstWaypoint) return;
 
@@ -302,7 +306,9 @@ async function update() {
                     calculatedArrival.depDist -= appliedDistance;
                 }
 
-                if (pilot.groundspeed >= 50) {
+                if (pilot.groundspeed >= 50 && (getKeyedValueFromSettings('map.navigraph.routeParsing.airportOverlay.dashedLine') || extendedPilot) && (
+                    !hideLineIfNoProcedure || kind === 'sids' || kind === 'stars' || i > 1 || extendedPilot
+                )) {
                     addFeature(`enroute-${ callsign }`, () => ({
                         geometry: turfGeometryToOl(greatCircle(coordinate, newCoordinate, { npoints: 16 })),
                         key: '',
@@ -319,7 +325,7 @@ async function update() {
                 firstWaypoint = true;
             };
 
-            for (let i = 0; i < waypoints.length; i++) {
+            for (i = 0; i < waypoints.length; i++) {
                 const waypoint = waypoints[i];
                 const nextWaypoint = waypoints[i + 1];
                 const nextCoordinate = nextWaypoint?.coordinate ?? [nextWaypoint?.airway?.value?.[2][0]?.[3], nextWaypoint?.airway?.value?.[2][0]?.[4]];